successfully implemented async

+++ src/sorting/sorting_methods.py
@@ -5,6 +5,8 @@ import shutil
 from PIL import Image
 from pathlib import Path
 
+from logger import logger
+
 from typing import Callable, Any, Optional
 
 SORTING_METHODS: dict[str, Callable[..., Any]] = {}
@@ -18,11 +20,19 @@ def register_sm(name: Optional[str] = None):
 @register_sm("move")
 def move(filter_group_names:list[str], image_path: Path, output_folder:Path):
     """makes (a) subfolder(s) in the output_folder named according to the filter_group_names(or just check if they exist and are writable if it already exists), moves copies of the original image in those folders and then deletes the original image folder"""
+    amount_skipped = 0
     for f in filter_group_names:
         fg_folder = output_folder / f
         fg_folder.mkdir(parents=True, exist_ok=True)
-        shutil.copy2(image_path,fg_folder)
-    os.remove(image_path)
+        if not (fg_folder / image_path.name).exists():
+            shutil.copy2(image_path,fg_folder)
+        else:
+            logger.warning(f"{(fg_folder / image_path.name)} already exists, skipping")
+            amount_skipped += 1
+    if not (amount_skipped == len(filter_group_names)):
+        os.remove(image_path)
+    else:
+        logger.warning(f"did not remove {image_path} because a file with the same name was everywhere it couldve been sorted to")
 
 @register_sm("link")
 def link(filter_group_names:list[str], image_path: Path, output_folder:Path):
@@ -30,7 +40,8 @@ def link(filter_group_names:list[str], image_path: Path, output_folder:Path):
     for f in filter_group_names:
         fg_folder = output_folder / f
         fg_folder.mkdir(parents=True, exist_ok=True)
-        os.symlink(image_path.resolve(), fg_folder / image_path.name)
+        if not (fg_folder / image_path.name).exists():
+            os.symlink(image_path.resolve(), fg_folder / image_path.name)
 
 @register_sm("tag")
 def tag(filter_group_names:list[str], image_path: Path, output_folder:Optional[Path] = None):
@@ -46,7 +57,10 @@ def name(filter_group_names:list[str], image_path: Path, output_folder:Optional[
     """changes the image file name to: `{filter_group_names(seperated with underscores)}{index of picture with the same filter_group_names combination}`"""
     new_name = "_".join(filter_group_names) + image_path.suffix
     new_path = image_path.parent / new_name
-    image_path.rename(new_path)
+    if not image_path.exists():
+        image_path.rename(new_path)
+    else:
+        logger.warning(f"could not rename '{image_path}' because the file already exists")
 
 @register_sm("json")
 def json(filter_group_names:list[str], image_path: Path, output_folder:Path):
